+++ src/Geta.Optimizely.ProductFeed/IProductFeedContentLoader.cs
@@ -2,12 +2,13 @@
 // Licensed under Apache-2.0. See the LICENSE file in the project root for more information
 
 using System.Collections.Generic;
+using System.Threading;
 using EPiServer.Commerce.Catalog.ContentTypes;
 
 namespace Geta.Optimizely.ProductFeed
 {
     public interface IProductFeedContentLoader
     {
-        ICollection<CatalogContentBase> LoadSourceData();
+        ICollection<CatalogContentBase> LoadSourceData(CancellationToken cancellationToken);
     }
 }