+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.0.0]
+
+* Extracted Google feed code into separate package (`Geta.Optimizely.ProductFeed.Google`)
+* Renamed package to `Geta.Optimizely.ProductFeed`
+
 ## [1.0.0]
 
 Initial Optimizely version of the package.